+++ tests/testthat/test-helpers_ggbetweenstats.R
@@ -44,7 +44,7 @@ testthat::test_that(
     # check data
     testthat::expect_equal(dim(pb_mean$data[[1]]), c(5L, 25L))
     testthat::expect_equal(dim(pb_mean$data[[2]]), c(5L, 12L))
-    testthat::expect_equal(dim(pb_mean$data[[3]]), c(4L, 15L))
+    # testthat::expect_equal(dim(pb_mean$data[[3]]), c(4L, 15L))
     testthat::expect_equal(
       pb_mean$data[[2]]$y,
       c(0.07925556, 0.62159750, 0.02155000, 0.14573118, 0.00762600),
@@ -83,7 +83,7 @@ testthat::test_that(
         x = cut,
         y = price,
         mean.ci = TRUE,
-        k = 3
+        k = 4
       )
 
     # check that dropped factor level is not retained
@@ -96,10 +96,10 @@ testthat::test_that(
     testthat::expect_identical(
       mean_dat$label,
       c(
-        "list(~italic(mu)==3819.580,CI[95*'%'](3140.804,4498.356))",
-        "list(~italic(mu)==4053.440,CI[95*'%'](3613.102,4493.778))",
-        "list(~italic(mu)==3928.250,CI[95*'%'](3607.033,4249.467))",
-        "list(~italic(mu)==4602.090,CI[95*'%'](4274.733,4929.447))"
+        "list(~italic(mu)==3819.5769,CI[95*'%']*'['*3140.8008,4498.3531*']')",
+        "list(~italic(mu)==4053.4397,CI[95*'%']*'['*3613.1020,4493.7773*']')",
+        "list(~italic(mu)==3928.2492,CI[95*'%']*'['*3607.0316,4249.4667*']')",
+        "list(~italic(mu)==4602.0879,CI[95*'%']*'['*4274.7306,4929.4452*']')"
       )
     )
 
@@ -118,13 +118,14 @@ testthat::test_that(
 
     # ggstatsplot output
     set.seed(123)
-    mean_dat2 <- ggstatsplot:::mean_labeller(
-      data = ggplot2::msleep,
-      x = vore,
-      y = brainwt,
-      mean.ci = TRUE,
-      k = 3
-    )
+    mean_dat2 <-
+      ggstatsplot:::mean_labeller(
+        data = ggplot2::msleep,
+        x = vore,
+        y = brainwt,
+        mean.ci = TRUE,
+        k = 3
+      )
 
     # when factor level contains NAs
     testthat::expect_equal(
@@ -173,6 +174,88 @@ testthat::test_that(
 
     testthat::expect_equal(dim(df2), c(83L, 13L))
     testthat::expect_equal(dim(dplyr::filter(df2, !is.na(outlier))), c(4L, 13L))
+  }
+)
+
+# `mean` pattern names ---------------------------------------------------
+
+testthat::test_that(
+  desc = "mean` pattern names ",
+  code = {
+    testthat::skip_on_cran()
+    # setup
+    set.seed(123)
+
+    # data
+    df <- structure(list(condition = structure(c(
+      1L, 1L, 1L, 1L, 1L, 1L,
+      1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 1L, 2L,
+      2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L,
+      2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 2L, 3L,
+      3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L,
+      3L, 3L, 3L, 3L, 3L, 3L, 3L, 3L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L,
+      4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L, 4L
+    ), .Label = c(
+      "0",
+      "1", "2", "3"
+    ), class = "factor"), size = c(
+      3, 1, 1, 2, 5, 4,
+      5, 3, 1, 1, 4, 3, 5, 4, 4, 4, 2, 2, 4, 5, 3, 3, 3, 3, 5, 1, 5,
+      5, 5, 1, 3, 4, 2, 1, 2, 1, 1, 3, 3, 1, 2, 3, 1, 4, 5, 5, 1, 5,
+      4, 5, 5, 1, 1, 4, 1, 2, 5, 1, 2, 2, 5, 3, 3, 4, 5, 3, 3, 3, 2,
+      1, 2, 4, 1, 1, 4, 4, 1, 2, NA, 3, 1, 4, 4, 2, 3, 4, 4, 4, 3,
+      5, 4, 2, 2, 5, 5, 5, 4, 1, 2, 5, 5
+    ), predict = c(
+      4, 4, 1, 1,
+      1, 4, 2, 4, 3, 2, 2, 3, 1, 1, 4, 3, 5, 2, 4, 2, 1, 5, 3, 3, 3,
+      3, 4, 2, 1, 1, 5, 2, 5, 3, 3, 3, 1, 5, 2, 3, 5, 2, 2, 5, 3, 2,
+      1, 4, 2, 2, 4, 4, 1, 4, 3, 3, 1, 1, 2, 3, 4, 4, 2, 5, 4, 3, 2,
+      3, 4, 4, 5, 2, 2, 4, 2, 2, 5, 4, NA, 1, 2, 3, 3, 5, 5, 5, 5,
+      1, 1, 1, 2, 1, 4, 2, 1, 2, 5, 3, 1, 4, 5
+    ), meaningful = c(
+      6,
+      5, 3, 3, 5, 4, 3, 2, 4, 6, 6, 4, 2, 2, 4, 5, 2, 5, 2, 4, 5, 1,
+      2, 7, 5, 7, 6, 3, 4, 4, 3, 7, 2, 2, 2, 4, 3, 3, 1, 6, 7, 1, 5,
+      1, 7, 4, 1, 2, 3, 4, 1, 1, 4, 1, 7, 3, 4, 7, 6, 6, 2, 5, 5, 6,
+      4, 3, 5, 6, 4, 1, 1, 2, 1, 4, 7, 5, 4, 6, NA, 5, 5, 6, 7, 4,
+      3, 7, 7, 5, 4, 3, 1, 5, 5, 1, 6, 1, 5, 2, 5, 2, 1
+    )), row.names = c(
+      NA,
+      -101L
+    ), class = c("tbl_df", "tbl", "data.frame"))
+
+    # another dataset
+    a <- data.frame(
+      mean.a = c(1.1, 0.9, 0.94, 1.58, 1.2, 1.4),
+      group = c("a", "a", "a", "b", "b", "b")
+    )
+
+    # summary
+    mean_data1 <-
+      ggstatsplot:::mean_labeller(
+        data = df,
+        x = condition,
+        y = meaningful,
+        messages = FALSE
+      )
+
+    # summary
+    mean_data2 <-
+      ggstatsplot:::mean_labeller(
+        data = a,
+        x = group,
+        y = mean.a,
+        messages = FALSE
+      )
+
+    testthat::expect_true(is.numeric(mean_data1$meaningful))
+    testthat::expect_true(is.numeric(mean_data2$mean.a))
+    testthat::expect_true(is.character(mean_data1$label))
+    testthat::expect_true(is.character(mean_data2$label))
+    testthat::expect_true(is.character(mean_data1$n_label))
+    testthat::expect_true(is.character(mean_data2$n_label))
+    testthat::expect_true(is.factor(mean_data1$condition))
+    testthat::expect_true(is.factor(mean_data2$group))
   }
 )